GUI changed
More natural feeling, less whitespace.
The idea here is to let it look a bit more like the mobile app, and in
the future have a static top bar instead of a changing one for every
page.

Two things:
1. Remaining likes etc didn't work for me, so removed. If this is a
mistake, let me know.
2. Undo swipe button is now to the left of the main photo. This is done
relative to the parent container. Better would be to have it relative to
the main photo so it moves along with window resizing.