Pull font-awesome from CDN rather than installing.
Using font-awesome as a dependency was creating additional build overhead.
Since we're already loading the Raleway font from Google's CDN, I didn't see a
reason not to load this font from a CDN as well.

This cuts down on the build logic, and also makes the migration to npm and away
from bower easier. Lastly it will cut down on diff churn as the fonts and css
were checked into the repository.